+++ CHANGELOG.md
@@ -12,6 +12,8 @@ omnibus-gitlab repository.
 - Upgrade libpng to 1.6.36 !2950
 - Fix invalid registry redirect url when using the default ports (jelhan) !2961
 - Update runit cookbook to v4.3.0 !2902
+- Turn on http to https redirection for Registry and Mattermost if LE
+  integration is used !2968
 
 11.7.0